+++ src/main/java/org/example/bpindividuallabjws/services/BlogpostService.java
@@ -1,6 +1,10 @@
 package org.example.bpindividuallabjws.services;
 
 import org.example.bpindividuallabjws.entities.Blogpost;
+import org.example.bpindividuallabjws.exceptions.BlankFieldException;
+import org.example.bpindividuallabjws.exceptions.EmptyObjectException;
+import org.example.bpindividuallabjws.exceptions.InvalidUserException;
+import org.example.bpindividuallabjws.exceptions.ResourceNotFoundException;
 import org.example.bpindividuallabjws.repositories.BlogpostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,22 +30,44 @@ public class BlogpostService implements BlogpostServiceInterface{
         @Override
         public Blogpost getSpecificBlogpost(Long id) {
                 Optional<Blogpost> retrievedBlogpost = blogpostRepository.findById(id);
-                return retrievedBlogpost.get();
+                if (retrievedBlogpost.isPresent()){
+                        return retrievedBlogpost.get();
+                } else {
+                        throw new ResourceNotFoundException("Blogpost", "ID", id);
+                }
         }
 
         @Override
         public Blogpost createBlogpost(Blogpost blogpost) {
-                return blogpostRepository.save(blogpost);
+                if (blogpost.getContent() == null){
+                        throw new EmptyObjectException("Blogpost", "content");
+                } else {
+                        if (blogpost.getCreator().isBlank() || blogpost.getContent().isEmpty()){
+                                blogpostRepository.save(blogpost);
+                                throw new BlankFieldException("blogpost", "content", "ID", blogpost.getId());
+                        }
+                        return blogpostRepository.save(blogpost);
+                }
         }
 
         @Override
         public Blogpost updateBlogpost(Blogpost blogpost) {
-                return blogpostRepository.save(blogpost);
+                Optional<Blogpost> blogpostToUpdate = blogpostRepository.findById(blogpost.getId());
+                if (blogpostToUpdate.isPresent()){
+                        return blogpostRepository.save(blogpost);
+                } else {
+                        throw new ResourceNotFoundException("Blogpost", "ID", blogpost.getId());
+                }
         }
 
         @Override
         public void deleteBlogpostByID(Long id) {
-                blogpostRepository.deleteById(id);
+                Optional<Blogpost> blogpostToDelete = blogpostRepository.findById(id);
+                if (blogpostToDelete.isPresent()){
+                        blogpostRepository.deleteById(id);
+                } else {
+                        throw new ResourceNotFoundException("Blogpost", "ID", id);
+                }
         }
 
         @Override
